Improve quick chat spam
Using quick chat spam is now suggested
