+++ Modules/ChatManager.cs
@@ -3,6 +3,7 @@ using Hazel;
 using System;
 using System.Security.Cryptography;
 using System.Text;
+using TOHE.Modules.Rpc;
 using TOHE.Roles.Impostor;
 using static TOHE.Options;
 using static TOHE.Translator;
@@ -160,91 +161,8 @@ namespace TOHE.Modules.ChatManager
 
         public static void SendQuickChatSpam()
         {
-            var firstAlivePlayer = Main.AllAlivePlayerControls.OrderBy(x => x.PlayerId).FirstOrDefault() ?? PlayerControl.LocalPlayer;
-            var title = "<color=#aaaaff>" + GetString("DefaultSystemMessageTitle") + "</color>";
-            var name = firstAlivePlayer?.Data?.PlayerName ?? "Error";
-
-            var writer = CustomRpcSender.Create("EzHacked_QuickChatSpamExploit", ExtendedPlayerControl.RpcSendOption);
-            writer.AutoStartRpc(firstAlivePlayer.NetId, (byte)RpcCalls.SetName);
-            writer.Write(firstAlivePlayer.Data.NetId);
-            writer.Write(title);
-            writer.EndRpc();
-
-            firstAlivePlayer.Data.PlayerName = title;
-
-            switch (quickChatSpamMode)
-            {
-                case QuickChatSpamMode.QuickChatSpam_Disabled:
-                    Logger.Info("QuickChatSpam disabled but trying to spam?", "SendQuickChatSpam");
-                    goto case QuickChatSpamMode.QuickChatSpam_Random20;
-                // Send as random 20 here
-                case QuickChatSpamMode.QuickChatSpam_Random20:
-                    var random = IRandom.Instance;
-                    var stringNamesValues = Enum.GetValues(typeof(StringNames)).Cast<StringNames>().ToArray();
-                    for (int i = 0; i < 25; i++)
-                    {
-                        var randomString = stringNamesValues[random.Next(stringNamesValues.Length)];
-                        writer.AutoStartRpc(firstAlivePlayer.NetId, (byte)RpcCalls.SendQuickChat);
-                        writer.Write((byte)QuickChatPhraseType.ComplexPhrase);
-                        writer.Write((uint)randomString);
-                        writer.Write((byte)0);
-                        writer.EndRpc();
-                        DestroyableSingleton<HudManager>.Instance.Chat.AddChat(firstAlivePlayer, GetString(randomString), false);
-                    }
-                    break;
-                case QuickChatSpamMode.QuickChatSpam_How2PlayNormal:
-                    foreach (var names in Main.how2playN)
-                    {
-                        writer.AutoStartRpc(firstAlivePlayer.NetId, (byte)RpcCalls.SendQuickChat);
-                        writer.Write((byte)QuickChatPhraseType.SimplePhrase);
-                        writer.Write((uint)names);
-                        writer.EndRpc();
-                        writer.AutoStartRpc(firstAlivePlayer.NetId, (byte)RpcCalls.SendQuickChat);
-                        writer.Write((byte)QuickChatPhraseType.SimplePhrase);
-                        writer.Write((uint)names);
-                        writer.EndRpc();
-                        DestroyableSingleton<HudManager>.Instance.Chat.AddChat(firstAlivePlayer, GetString(names), false);
-                        DestroyableSingleton<HudManager>.Instance.Chat.AddChat(firstAlivePlayer, GetString(names), false);
-                    }
-                    break;
-                case QuickChatSpamMode.QuickChatSpam_How2PlayHidenSeek:
-                    foreach (var names in Main.how2playHnS)
-                    {
-                        writer.AutoStartRpc(firstAlivePlayer.NetId, (byte)RpcCalls.SendQuickChat);
-                        writer.Write((byte)QuickChatPhraseType.SimplePhrase);
-                        writer.Write((uint)names);
-                        writer.EndRpc();
-                        writer.AutoStartRpc(firstAlivePlayer.NetId, (byte)RpcCalls.SendQuickChat);
-                        writer.Write((byte)QuickChatPhraseType.SimplePhrase);
-                        writer.Write((uint)names);
-                        writer.EndRpc();
-                        DestroyableSingleton<HudManager>.Instance.Chat.AddChat(firstAlivePlayer, GetString(names), false);
-                        DestroyableSingleton<HudManager>.Instance.Chat.AddChat(firstAlivePlayer, GetString(names), false);
-                    }
-                    break;
-                case QuickChatSpamMode.QuickChatSpam_EzHacked:
-                    foreach (var names in Main.how2playEzHacked)
-                    {
-                        writer.AutoStartRpc(firstAlivePlayer.NetId, (byte)RpcCalls.SendQuickChat);
-                        writer.Write((byte)QuickChatPhraseType.SimplePhrase);
-                        writer.Write((uint)names);
-                        writer.EndRpc();
-                        writer.AutoStartRpc(firstAlivePlayer.NetId, (byte)RpcCalls.SendQuickChat);
-                        writer.Write((byte)QuickChatPhraseType.SimplePhrase);
-                        writer.Write((uint)names);
-                        writer.EndRpc();
-                        DestroyableSingleton<HudManager>.Instance.Chat.AddChat(firstAlivePlayer, GetString(names), false);
-                        DestroyableSingleton<HudManager>.Instance.Chat.AddChat(firstAlivePlayer, GetString(names), false);
-                    }
-                    break;
-            }
-            writer.AutoStartRpc(firstAlivePlayer.NetId, (byte)RpcCalls.SetName);
-            writer.Write(firstAlivePlayer.Data.NetId);
-            writer.Write(name);
-            writer.EndRpc();
-
-            firstAlivePlayer.Data.PlayerName = name;
-            writer.SendMessage();
+            var message = new RpcQuickChatSpam();
+            RpcUtils.LateBroadcastReliableMessage(message);
         }
         public static void SendPreviousMessagesToAll()
         {
