+++ Resources/Lang/en_US.json
@@ -4084,6 +4084,7 @@
   "QuickChatSpam_How2PlayHidenSeek": "How2Play HnS",
   "QuickChatSpam_Random20": "Random Strings",
   "QuickChatSpam_EzHacked": "Ducked Ezly",
+  "QuickChatSpam_Empty": "Empty String",
   "CrossLanguageGetRole": "Get role names from different languages for commands",
 
   "BypassRateLimitAC": "Experimental Changes to Bypass RateLimit Anticheat",
